+++ Mods ChoGGi/Fix Buildings Broken Down And No Repair/metadata.lua
@@ -1,0 +1,18 @@
+return PlaceObj("ModDef", {
+	"title", "Fix: Buildings Broken Down And No Repair v0.1",
+	"version", 1,
+	"saved", 1545566400,
+	"image", "Preview.png",
+	"id", "ChoGGi_FixBuildingsBrokenDownAndNoRepair",
+	"steam_id", "1599190080",
+	"author", "ChoGGi",
+	"lua_revision", LuaRevision,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[The Long Winter event seems to be responsible.
+
+If you have broken down buildings the drones won't repair.
+
+Any buildings affected by this issue will need to be repaired with 000.1 resource after the fix happens.]],
+})